Example of running multiple similar test cases
In this example, how a same test case, using various set of values can be written in a compact manner. Please look at the output file for a better idea, in the output, the last test case was made failed to show an example of failure of one test case from multiple cases.